@@ -1,9 +1,9 @@
 {
   "name": "vnopts",
+  "type": "module",
   "version": "1.0.2",
   "description": "validate and normalize options",
-  "main": "lib/index.js",
-  "types": "lib/index.d.ts",
+  "exports": "./lib/index.js",
   "repository": "https://github.com/ikatyang/vnopts",
   "homepage": "https://github.com/ikatyang/vnopts#readme",
   "author": {
@@ -13,39 +13,33 @@
   },
   "license": "MIT",
   "scripts": {
-    "prepublish": "yarn run build",
-    "lint": "tslint -p ./tsconfig.json --type-check",
-    "test": "jest",
-    "prebuild": "rm -rf ./lib",
+    "prepack": "pnpm run build",
+    "lint": "prettier --check .",
+    "test": "vitest",
     "build": "tsc -p ./tsconfig.build.json",
     "release": "standard-version"
   },
   "dependencies": {
-    "chalk": "^2.4.1",
-    "leven": "^2.1.0",
-    "tslib": "^1.9.3"
+    "chalk": "^5.3.0",
+    "leven": "^4.0.0",
+    "tslib": "^2.6.0"
   },
   "devDependencies": {
-    "@types/jest": "23.3.3",
-    "@types/leven": "2.1.1",
-    "@types/node": "6.0.118",
-    "has-ansi": "3.0.0",
-    "jest": "23.6.0",
-    "jest-snapshot-serializer-ansi": "1.0.0",
-    "prettier": "1.14.3",
-    "prettier-config-ikatyang": "1.1.1",
-    "standard-version": "4.4.0",
-    "strip-ansi": "4.0.0",
-    "ts-jest": "23.10.3",
-    "tslint": "5.11.0",
-    "tslint-config-prettier": "1.15.0",
-    "tslint-plugin-prettier": "2.0.0",
-    "typescript": "3.0.3"
+    "@types/node": "18.16.19",
+    "@vitest/coverage-v8": "0.33.0",
+    "jest-snapshot-serializer-ansi": "2.1.0",
+    "prettier": "3.0.0",
+    "standard-version": "9.5.0",
+    "typescript": "5.1.6",
+    "vite": "4.4.4",
+    "vitest": "0.33.0"
   },
   "engines": {
-    "node": ">= 6"
+    "node": ">=16"
   },
   "files": [
-    "/lib/**/*"
-  ]
+    "/lib/**/*",
+    "/CHANGELOG.md"
+  ],
+  "packageManager": "pnpm@8.6.6"
 }
