@@ -1,6 +1,6 @@
-import chalk from 'chalk';
-import leven = require('leven');
-import { UnknownHandler } from '../../types';
+import chalk from 'chalk'
+import leven from 'leven'
+import { UnknownHandler } from '../../types.js'
 
 export const levenUnknownHandler: UnknownHandler = (
   key,
@@ -9,15 +9,15 @@ export const levenUnknownHandler: UnknownHandler = (
 ) => {
   const messages = [
     `Ignored unknown option ${chalk.yellow(descriptor.pair({ key, value }))}.`,
-  ];
+  ]
 
   const suggestion = Object.keys(schemas)
     .sort()
-    .find(knownKey => leven(key, knownKey) < 3);
+    .find(knownKey => leven(key, knownKey) < 3)
 
   if (suggestion) {
-    messages.push(`Did you mean ${chalk.blue(descriptor.key(suggestion))}?`);
+    messages.push(`Did you mean ${chalk.blue(descriptor.key(suggestion))}?`)
   }
 
-  logger.warn(messages.join(' '));
-};
+  logger.warn(messages.join(' '))
+}
