@@ -1,2 +1,2 @@
-export const VALUE_NOT_EXIST = Symbol.for('vnopts.VALUE_NOT_EXIST');
-export const VALUE_UNCHANGED = Symbol.for('vnopts.VALUE_UNCHANGED');
+export const VALUE_NOT_EXIST = Symbol.for('vnopts.VALUE_NOT_EXIST')
+export const VALUE_UNCHANGED = Symbol.for('vnopts.VALUE_UNCHANGED')
