@@ -1,3 +1,3 @@
-export * from './deprecated';
-export * from './invalid';
-export * from './unknown';
+export * from './deprecated/index.js'
+export * from './invalid/index.js'
+export * from './unknown/index.js'
