@@ -1,25 +1,24 @@
-import chalk from 'chalk';
-import { VALUE_NOT_EXIST } from '../../constants';
+import chalk from 'chalk'
+import { VALUE_NOT_EXIST } from '../../constants.js'
 import {
   Descriptor,
   InvalidHandler,
   NormalizedExpectedResult,
   OptionKey,
   OptionValue,
-} from '../../types';
+} from '../../types.js'
 
-const INDENTATION = ' '.repeat(2);
+const INDENTATION = ' '.repeat(2)
 
 export const commonInvalidHandler: InvalidHandler = (key, value, utils) => {
   const { text, list } = utils.normalizeExpectedResult(
     utils.schemas[key].expected(utils),
-  );
+  )
 
-  const descriptions: string[] = [];
+  const descriptions: string[] = []
 
-  // istanbul ignore else
   if (text) {
-    descriptions.push(getDescription(key, value, text, utils.descriptor));
+    descriptions.push(getDescription(key, value, text, utils.descriptor))
   }
 
   if (list) {
@@ -31,11 +30,11 @@ export const commonInvalidHandler: InvalidHandler = (key, value, utils) => {
           ),
         )
         .join('\n'),
-    );
+    )
   }
 
-  return chooseDescription(descriptions, utils.loggerPrintWidth);
-};
+  return chooseDescription(descriptions, utils.loggerPrintWidth)
+}
 
 function getDescription(
   key: OptionKey,
@@ -51,17 +50,17 @@ function getDescription(
         ? chalk.gray('nothing')
         : chalk.red(descriptor.value(value))
     }.`,
-  ].join(' ');
+  ].join(' ')
 }
 
 function getListDescription(
   { text, list }: NormalizedExpectedResult,
   printWidth: number,
 ): string {
-  const descriptions: string[] = [];
+  const descriptions: string[] = []
 
   if (text) {
-    descriptions.push(`- ${chalk.blue(text)}`);
+    descriptions.push(`- ${chalk.blue(text)}`)
   }
 
   if (list) {
@@ -76,23 +75,23 @@ function getListDescription(
           ),
         )
         .join('\n'),
-    );
+    )
   }
 
-  return chooseDescription(descriptions, printWidth);
+  return chooseDescription(descriptions, printWidth)
 }
 
 function chooseDescription(descriptions: string[], printWidth: number) {
   if (descriptions.length === 1) {
-    return descriptions[0];
+    return descriptions[0]
   }
 
-  const [firstDescription, secondDescription] = descriptions;
+  const [firstDescription, secondDescription] = descriptions
   const [firstWidth, secondWidth] = descriptions.map(
     description => description.split('\n', 1)[0].length,
-  );
+  )
 
   return firstWidth > printWidth && firstWidth > secondWidth
     ? secondDescription
-    : firstDescription;
+    : firstDescription
 }
