@@ -1,4 +1,4 @@
-import { VALUE_UNCHANGED } from './constants';
+import { VALUE_UNCHANGED } from './constants.js'
 import {
   DefaultResult,
   DeprecatedResult,
@@ -8,63 +8,51 @@ import {
   RedirectResult,
   Utils,
   ValidateResult,
-} from './types';
+} from './types.js'
 
 export interface SchemaParameters<$Value> extends SchemaHandlers<$Value> {
-  name: string;
+  name: string
 }
 
 export interface SchemaHandlers<$Value> {
   default?:
     | DefaultResult<$Value>
-    | ((schema: Schema<$Value>, utils: Utils) => DefaultResult<$Value>);
+    | ((schema: Schema<$Value>, utils: Utils) => DefaultResult<$Value>)
   expected?:
     | ExpectedResult
-    | ((schema: Schema<$Value>, utils: Utils) => ExpectedResult);
+    | ((schema: Schema<$Value>, utils: Utils) => ExpectedResult)
   validate?:
     | ValidateResult
-    | ((
-        value: unknown,
-        schema: Schema<$Value>,
-        utils: Utils,
-      ) => ValidateResult);
+    | ((value: unknown, schema: Schema<$Value>, utils: Utils) => ValidateResult)
   deprecated?:
     | DeprecatedResult<$Value>
     | ((
         value: $Value,
         schema: Schema<$Value>,
         utils: Utils,
-      ) => DeprecatedResult<$Value>);
+      ) => DeprecatedResult<$Value>)
   forward?:
     | ForwardResult<$Value>
     | ((
         value: $Value,
         schema: Schema<$Value>,
         utils: Utils,
-      ) => ForwardResult<$Value>);
+      ) => ForwardResult<$Value>)
   redirect?:
     | RedirectResult<$Value>
     | ((
         value: $Value,
         schema: Schema<$Value>,
         utils: Utils,
-      ) => RedirectResult<$Value>);
+      ) => RedirectResult<$Value>)
   overlap?: (
     currentValue: $Value,
     newValue: $Value,
     schema: Schema<$Value>,
     utils: Utils,
-  ) => $Value;
-  preprocess?: (
-    value: unknown,
-    schema: Schema<$Value>,
-    utils: Utils,
-  ) => unknown;
-  postprocess?: (
-    value: $Value,
-    schema: Schema<$Value>,
-    utils: Utils,
-  ) => unknown;
+  ) => $Value
+  preprocess?: (value: unknown, schema: Schema<$Value>, utils: Utils) => unknown
+  postprocess?: (value: $Value, schema: Schema<$Value>, utils: Utils) => unknown
 }
 
 const HANDLER_KEYS: Array<keyof SchemaHandlers<any>> = [
@@ -77,75 +65,80 @@ const HANDLER_KEYS: Array<keyof SchemaHandlers<any>> = [
   'overlap',
   'preprocess',
   'postprocess',
-];
+]
 
 export function createSchema<
   $Schema extends Schema<any>,
-  $Parameters extends $Schema['_parametersType']
+  $Parameters extends $Schema['_parametersType'],
 >(
   SchemaConstructor: new (parameters: $Parameters) => $Schema,
   parameters: $Parameters,
 ) {
-  const schema = new SchemaConstructor(parameters);
-  const subSchema = Object.create(schema) as $Schema;
+  const schema = new SchemaConstructor(parameters)
+  const subSchema = Object.create(schema) as $Schema
 
   for (const handlerKey of HANDLER_KEYS) {
     if (handlerKey in parameters) {
+      // @ts-ignore
       subSchema[handlerKey] = normalizeHandler(
         parameters[handlerKey],
         schema,
         Schema.prototype[handlerKey].length,
-      );
+      )
     }
   }
 
-  return subSchema;
+  return subSchema
 }
 
 export abstract class Schema<
   $Value extends OptionValue,
-  $Parameters extends SchemaHandlers<$Value> = SchemaHandlers<$Value>
+  $Parameters extends SchemaHandlers<$Value> = SchemaHandlers<$Value>,
 > {
   public static create<$Schema extends Schema<any>>(
     parameters: $Schema['_parametersType'],
   ): $Schema {
     // @ts-ignore: https://github.com/Microsoft/TypeScript/issues/5863
-    return createSchema(this, parameters);
+    return createSchema(this, parameters)
   }
 
-  public name: string;
+  public name: string
 
-  public _valueType!: $Value;
-  public _parametersType!: $Parameters;
+  public _valueType!: $Value
+  public _parametersType!: $Parameters
 
   constructor(parameters: SchemaParameters<$Value>) {
-    this.name = parameters.name;
+    this.name = parameters.name
   }
 
   public default(_utils: Utils): DefaultResult<$Value> {
-    return undefined;
+    return undefined
   }
 
-  // istanbul ignore next: this is actually an abstract method but we need a placeholder to get `function.length`
+  // this is actually an abstract method but we need a placeholder to get `function.length`
+  /* c8 ignore start */
   public expected(_utils: Utils): ExpectedResult {
-    return 'nothing';
+    return 'nothing'
   }
+  /* c8 ignore stop */
 
-  // istanbul ignore next: this is actually an abstract method but we need a placeholder to get `function.length`
+  // this is actually an abstract method but we need a placeholder to get `function.length`
+  /* c8 ignore start */
   public validate(_value: unknown, _utils: Utils): ValidateResult {
-    return false;
+    return false
   }
+  /* c8 ignore stop */
 
   public deprecated(_value: $Value, _utils: Utils): DeprecatedResult<$Value> {
-    return false;
+    return false
   }
 
   public forward(_value: $Value, _utils: Utils): ForwardResult<$Value> {
-    return undefined;
+    return undefined
   }
 
   public redirect(_value: $Value, _utils: Utils): RedirectResult<$Value> {
-    return undefined;
+    return undefined
   }
 
   public overlap(
@@ -153,15 +146,15 @@ export abstract class Schema<
     _newValue: $Value,
     _utils: Utils,
   ): $Value {
-    return currentValue;
+    return currentValue
   }
 
   public preprocess(value: unknown, _utils: Utils): any {
-    return value;
+    return value
   }
 
   public postprocess(_value: $Value, _utils: Utils): any {
-    return VALUE_UNCHANGED;
+    return VALUE_UNCHANGED
   }
 }
 
@@ -172,10 +165,10 @@ function normalizeHandler<$Result>(
 ) {
   return typeof handler === 'function'
     ? (...args: any[]) =>
-        handler(
+        (handler as (...args: any[]) => $Result)(
           ...args.slice(0, handlerArgumentsLength - 1),
           superSchema,
           ...args.slice(handlerArgumentsLength - 1),
         )
-    : () => handler;
+    : () => handler
 }
