@@ -1,4 +1,4 @@
-import { Schema, SchemaHandlers } from '../schema';
+import { Schema, SchemaHandlers } from '../schema.js'
 import {
   DeprecatedResult,
   ExpectedResult,
@@ -7,33 +7,33 @@ import {
   RedirectResult,
   Utils,
   ValidateResult,
-} from '../types';
+} from '../types.js'
 
 interface ArraySchemaParameters<$ValueSchema extends Schema<any>>
   extends SchemaHandlers<Array<$ValueSchema['_valueType']>> {
-  name?: string;
-  valueSchema: $ValueSchema;
+  name?: string
+  valueSchema: $ValueSchema
 }
 
 export class ArraySchema<$ValueSchema extends Schema<any>> extends Schema<
   Array<$ValueSchema['_valueType']>,
   ArraySchemaParameters<$ValueSchema>
 > {
-  private _valueSchema: $ValueSchema;
+  private _valueSchema: $ValueSchema
 
   constructor({
     valueSchema,
     name = valueSchema.name,
     ...handlers
   }: ArraySchemaParameters<$ValueSchema>) {
-    super({ ...handlers, name });
-    this._valueSchema = valueSchema;
+    super({ ...handlers, name })
+    this._valueSchema = valueSchema
   }
 
   public expected(utils: Utils): ExpectedResult {
     const { text, list } = utils.normalizeExpectedResult(
       this._valueSchema.expected(utils),
-    );
+    )
 
     return {
       text: text && `an array of ${text}`,
@@ -41,104 +41,104 @@ export class ArraySchema<$ValueSchema extends Schema<any>> extends Schema<
         title: `an array of the following values`,
         values: [{ list }],
       },
-    };
+    }
   }
 
   public validate(value: unknown, utils: Utils): ValidateResult {
     if (!Array.isArray(value)) {
-      return false;
+      return false
     }
 
-    const invalidValues: Array<unknown> = [];
+    const invalidValues: Array<unknown> = []
 
     for (const subValue of value) {
       const subValidateResult = utils.normalizeValidateResult(
         this._valueSchema.validate(subValue, utils),
         subValue,
-      );
+      )
       if (subValidateResult !== true) {
-        invalidValues.push(subValidateResult.value);
+        invalidValues.push(subValidateResult.value)
       }
     }
 
-    return invalidValues.length === 0 ? true : { value: invalidValues };
+    return invalidValues.length === 0 ? true : { value: invalidValues }
   }
 
   public deprecated(
     value: this['_valueType'],
     utils: Utils,
   ): DeprecatedResult<this['_valueType']> {
-    const deprecatedResult: DeprecatedResult<this['_valueType']> = [];
+    const deprecatedResult: DeprecatedResult<this['_valueType']> = []
 
     for (const subValue of value) {
       const subDeprecatedResult = utils.normalizeDeprecatedResult(
         this._valueSchema.deprecated(subValue, utils),
         subValue,
-      );
+      )
 
       if (subDeprecatedResult !== false) {
         deprecatedResult.push(
           ...subDeprecatedResult.map(({ value: deprecatedValue }) => ({
             value: [deprecatedValue],
           })),
-        );
+        )
       }
     }
 
-    return deprecatedResult;
+    return deprecatedResult
   }
 
   public forward(
     value: this['_valueType'],
     utils: Utils,
   ): ForwardResult<this['_valueType']> {
-    const forwardResult: ForwardResult<this['_valueType']> = [];
+    const forwardResult: ForwardResult<this['_valueType']> = []
 
     for (const subValue of value) {
       const subForwardResult = utils.normalizeForwardResult(
         this._valueSchema.forward(subValue, utils),
         subValue,
-      );
+      )
 
-      forwardResult.push(...subForwardResult.map(wrapTransferResult));
+      forwardResult.push(...subForwardResult.map(wrapTransferResult))
     }
 
-    return forwardResult;
+    return forwardResult
   }
 
   public redirect(
     value: this['_valueType'],
     utils: Utils,
   ): RedirectResult<this['_valueType']> {
-    const remain: this['_valueType'] = [];
-    const redirect: ForwardResult<this['_valueType']> = [];
+    const remain: this['_valueType'] = []
+    const redirect: ForwardResult<this['_valueType']> = []
 
     for (const subValue of value) {
       const subRedirectResult = utils.normalizeRedirectResult(
         this._valueSchema.redirect(subValue, utils),
         subValue,
-      );
+      )
 
       if ('remain' in subRedirectResult) {
-        remain.push(subRedirectResult.remain!);
+        remain.push(subRedirectResult.remain!)
       }
 
-      redirect.push(...subRedirectResult.redirect.map(wrapTransferResult));
+      redirect.push(...subRedirectResult.redirect.map(wrapTransferResult))
     }
 
-    return remain.length === 0 ? { redirect } : { redirect, remain };
+    return remain.length === 0 ? { redirect } : { redirect, remain }
   }
 
   public overlap(
     currentValue: this['_valueType'],
     newValue: this['_valueType'],
   ): this['_valueType'] {
-    return currentValue.concat(newValue);
+    return currentValue.concat(newValue)
   }
 }
 
 function wrapTransferResult<$Value>(
   { from, to }: NormalizedTransferResult<$Value>, //
 ): NormalizedTransferResult<$Value[]> {
-  return { from: [from], to };
+  return { from: [from], to }
 }
