@@ -1,11 +1,11 @@
-import { Schema, SchemaParameters } from '../schema';
-import { ValidateResult } from '../types';
+import { Schema, SchemaParameters } from '../schema.js'
+import { ValidateResult } from '../types.js'
 
 export class BooleanSchema extends Schema<boolean, SchemaParameters<boolean>> {
   public expected() {
-    return 'true or false';
+    return 'true or false'
   }
   public validate(value: unknown): ValidateResult {
-    return typeof value === 'boolean';
+    return typeof value === 'boolean'
   }
 }
