@@ -1,8 +1,8 @@
-export * from './alias';
-export * from './any';
-export * from './array';
-export * from './boolean';
-export * from './choice';
-export * from './integer';
-export * from './number';
-export * from './string';
+export * from './alias.js'
+export * from './any.js'
+export * from './array.js'
+export * from './boolean.js'
+export * from './choice.js'
+export * from './integer.js'
+export * from './number.js'
+export * from './string.js'
