@@ -1,11 +1,11 @@
-import { Schema, SchemaParameters } from '../schema';
-import { ValidateResult } from '../types';
+import { Schema, SchemaParameters } from '../schema.js'
+import { ValidateResult } from '../types.js'
 
 export class StringSchema extends Schema<string, SchemaParameters<string>> {
   public expected() {
-    return 'a string';
+    return 'a string'
   }
   public validate(value: unknown): ValidateResult {
-    return typeof value === 'string';
+    return typeof value === 'string'
   }
 }
