@@ -1,5 +1,5 @@
-import { VALUE_UNCHANGED } from './constants';
-import { Schema } from './schema';
+import { VALUE_UNCHANGED } from './constants.js'
+import { Schema } from './schema.js'
 import {
   normalizeDefaultResult,
   normalizeDeprecatedResult,
@@ -8,137 +8,137 @@ import {
   normalizeRedirectResult,
   normalizeValidateResult,
   NotEmptyArray,
-} from './utils';
+} from './utils.js'
 
 export interface Utils {
-  logger: Logger;
-  loggerPrintWidth: number;
-  descriptor: Descriptor;
-  schemas: Record<string, Schema<any>>;
-  normalizeDefaultResult: typeof normalizeDefaultResult;
-  normalizeExpectedResult: typeof normalizeExpectedResult;
-  normalizeDeprecatedResult: typeof normalizeDeprecatedResult;
-  normalizeForwardResult: typeof normalizeForwardResult;
-  normalizeRedirectResult: typeof normalizeRedirectResult;
-  normalizeValidateResult: typeof normalizeValidateResult;
+  logger: Logger
+  loggerPrintWidth: number
+  descriptor: Descriptor
+  schemas: Record<string, Schema<any>>
+  normalizeDefaultResult: typeof normalizeDefaultResult
+  normalizeExpectedResult: typeof normalizeExpectedResult
+  normalizeDeprecatedResult: typeof normalizeDeprecatedResult
+  normalizeForwardResult: typeof normalizeForwardResult
+  normalizeRedirectResult: typeof normalizeRedirectResult
+  normalizeValidateResult: typeof normalizeValidateResult
 }
 
 export interface Logger {
-  warn(message: string): void;
+  warn(message: string): void
 }
 
 export interface Descriptor {
-  key: (key: OptionKey) => string;
-  value: (value: OptionValue) => string;
-  pair: (pair: OptionPair) => string;
+  key: (key: OptionKey) => string
+  value: (value: OptionValue) => string
+  pair: (pair: OptionPair) => string
 }
 
 export type UnknownHandler = (
   key: OptionKey,
   value: OptionValue,
   utils: Utils,
-) => void | Options;
+) => void | Options
 
 export type InvalidHandler = (
   key: OptionKey,
   value: OptionValue,
   utils: Utils,
-) => string | Error;
+) => string | Error
 export type NormalizedInvalidHandler = (
   key: OptionKey,
   value: OptionValue,
   utils: Utils,
-) => Error;
+) => Error
 
 export type DeprecatedHandler = (
   keyOrPair: OptionKey | OptionPair,
   redirectTo: undefined | TransferTo,
   utils: Utils,
-) => string;
+) => string
 
-export type IdentifyMissing = (key: string, options: Options) => boolean;
-export type IdentifyRequired = (key: string) => boolean;
+export type IdentifyMissing = (key: string, options: Options) => boolean
+export type IdentifyRequired = (key: string) => boolean
 
-export type Preprocess = (options: Options, utils: Utils) => Options;
-export type Postprocess = (options: Options, utils: Utils) => PostprocessResult;
+export type Preprocess = (options: Options, utils: Utils) => Options
+export type Postprocess = (options: Options, utils: Utils) => PostprocessResult
 export type PostprocessResult =
   | typeof VALUE_UNCHANGED
-  | { delete?: string[]; override?: Options };
+  | { delete?: string[]; override?: Options }
 
-export type OptionKey = string;
-export type OptionValue = any;
+export type OptionKey = string
+export type OptionValue = any
 export interface OptionPair {
-  key: OptionKey;
-  value: OptionValue;
+  key: OptionKey
+  value: OptionValue
 }
 export interface Options {
-  [key: string]: OptionValue;
+  [key: string]: OptionValue
 }
 
-export type TransferTo = OptionKey | OptionPair;
+export type TransferTo = OptionKey | OptionPair
 export type TransferResult<$Value> =
   | TransferTo
   | {
-      from?: $Value;
-      to: TransferTo;
-    };
+      from?: $Value
+      to: TransferTo
+    }
 export interface NormalizedTransferResult<$Value> {
-  from: $Value;
-  to: TransferTo;
+  from: $Value
+  to: TransferTo
 }
 
 export type ForwardResult<$Value> =
   | undefined
   | TransferResult<$Value>
-  | Array<TransferResult<$Value>>;
+  | Array<TransferResult<$Value>>
 export type NormalizedForwardResult<$Value> = Array<
   NormalizedTransferResult<$Value>
->;
+>
 
 export type RedirectResult<$Value> =
   | ForwardResult<$Value>
   | {
-      remain?: $Value;
-      redirect: ForwardResult<$Value>;
-    };
+      remain?: $Value
+      redirect: ForwardResult<$Value>
+    }
 export interface NormalizedRedirectResult<$Value> {
-  remain?: $Value;
-  redirect: NormalizedForwardResult<$Value>;
+  remain?: $Value
+  redirect: NormalizedForwardResult<$Value>
 }
 
-export type ValidateResult = boolean | { value: unknown };
-export type NormalizedValidateResult = true | { value: unknown };
+export type ValidateResult = boolean | { value: unknown }
+export type NormalizedValidateResult = true | { value: unknown }
 
 export type DeprecatedResult<$Value> =
   | boolean
   | { value: $Value }
-  | Array<{ value: $Value }>;
+  | Array<{ value: $Value }>
 export type NormalizedDeprecatedResult<$Value> =
   | false
-  | NotEmptyArray<{ value: $Value }>;
+  | NotEmptyArray<{ value: $Value }>
 export type NormalizedDeprecatedResultWithTrue<$Value> =
   | true
-  | NormalizedDeprecatedResult<$Value>;
+  | NormalizedDeprecatedResult<$Value>
 
 export type ExpectedResult =
   | string
   | {
       // at least one of the following field exists
-      text?: string;
+      text?: string
       list?: {
-        title: string;
-        values: ExpectedResult[];
-      };
-    };
+        title: string
+        values: ExpectedResult[]
+      }
+    }
 export interface NormalizedExpectedResult {
-  text?: string;
+  text?: string
   list?: {
-    title: string;
-    values: NormalizedExpectedResult[];
-  };
+    title: string
+    values: NormalizedExpectedResult[]
+  }
 }
 
-export type DefaultResult<$Value> = undefined | { value?: $Value };
+export type DefaultResult<$Value> = undefined | { value?: $Value }
 export interface NormalizedDefaultResult<$Value> {
-  value?: $Value;
+  value?: $Value
 }
