@@ -1,11 +1,12 @@
-import { comparePrimitive, normalizeDeprecatedResult } from './utils';
+import { expect, test } from 'vitest'
+import { comparePrimitive, normalizeDeprecatedResult } from './utils.js'
 
 test(comparePrimitive.name, () => {
   expect([1, 1, true, false, 2, 'string'].sort(comparePrimitive)) //
-    .toEqual([false, true, 1, 1, 2, 'string']);
-});
+    .toEqual([false, true, 1, 1, 2, 'string'])
+})
 
 test(normalizeDeprecatedResult.name, () => {
-  const value = 'hello';
-  expect(normalizeDeprecatedResult(true, value)).toEqual([{ value }]);
-});
+  const value = 'hello'
+  expect(normalizeDeprecatedResult(true, value)).toEqual([{ value }])
+})
