@@ -17,87 +17,89 @@ import {
   RedirectResult,
   TransferResult,
   ValidateResult,
-} from './types';
+} from './types.js'
 
-export type NotEmptyArray<T> = [T, ...T[]];
+export type NotEmptyArray<T> = [T, ...T[]]
 
 export type RecordFromUnion<T extends Options, K extends keyof T> = {
   [X in T[K]]: Extract<T, Record<K, X>>
-};
+}
 
 export function recordFromArray<
   T extends Record<string, any>,
-  K extends keyof T
+  K extends keyof T & string,
 >(array: T[], mainKey: K): RecordFromUnion<T, K> {
-  const record = Object.create(null) as Partial<RecordFromUnion<T, K>>;
+  const record = Object.create(null) as Partial<RecordFromUnion<T, K>>
 
   for (const value of array) {
-    const key = value[mainKey];
+    const key = value[mainKey]
 
-    // istanbul ignore next
+    /* c8 ignore start */
     if (record[key]) {
-      throw new Error(`Duplicate ${mainKey} ${JSON.stringify(key)}`);
+      throw new Error(`Duplicate ${mainKey} ${JSON.stringify(key)}`)
     }
+    /* c8 ignore stop */
 
     // @ts-ignore
-    record[key] = value;
+    record[key] = value
   }
 
-  return record as RecordFromUnion<T, K>;
+  return record as RecordFromUnion<T, K>
 }
 
-export function mapFromArray<T extends Record<string, any>, K extends keyof T>(
-  array: T[],
-  mainKey: K,
-): Map<T[K], T> {
-  const map = new Map<T[K], T>();
+export function mapFromArray<
+  T extends Record<string, any>,
+  K extends keyof T & string,
+>(array: T[], mainKey: K): Map<T[K], T> {
+  const map = new Map<T[K], T>()
 
   for (const value of array) {
-    const key = value[mainKey];
+    const key = value[mainKey]
 
-    // istanbul ignore next
+    /* c8 ignore start */
     if (map.has(key)) {
-      throw new Error(`Duplicate ${mainKey} ${JSON.stringify(key)}`);
+      throw new Error(`Duplicate ${mainKey} ${JSON.stringify(key)}`)
     }
+    /* c8 ignore stop */
 
-    map.set(key, value);
+    map.set(key, value)
   }
 
-  return map;
+  return map
 }
 
 export function createAutoChecklist() {
-  const map: { [idString: string]: boolean } = Object.create(null);
+  const map: { [idString: string]: boolean } = Object.create(null)
   return (id: any) => {
-    const idString = JSON.stringify(id);
+    const idString = JSON.stringify(id)
     if (map[idString]) {
-      return true;
+      return true
     }
-    map[idString] = true;
-    return false;
-  };
+    map[idString] = true
+    return false
+  }
 }
 
 export function partition<T>(
   array: T[],
   predicate: (x: T) => boolean,
 ): [T[], T[]] {
-  const trueArray: T[] = [];
-  const falseArray: T[] = [];
+  const trueArray: T[] = []
+  const falseArray: T[] = []
 
   for (const value of array) {
     if (predicate(value)) {
-      trueArray.push(value);
+      trueArray.push(value)
     } else {
-      falseArray.push(value);
+      falseArray.push(value)
     }
   }
 
-  return [trueArray, falseArray];
+  return [trueArray, falseArray]
 }
 
 export function isInt(value: number) {
-  return value === Math.floor(value);
+  return value === Math.floor(value)
 }
 
 export function comparePrimitive(
@@ -105,11 +107,11 @@ export function comparePrimitive(
   b: undefined | null | boolean | number | string,
 ): number {
   if (a === b) {
-    return 0;
+    return 0
   }
 
-  const typeofA = typeof a;
-  const typeofB = typeof b;
+  const typeofA = typeof a
+  const typeofB = typeof b
 
   const orders = [
     'undefined',
@@ -117,51 +119,51 @@ export function comparePrimitive(
     'boolean',
     'number',
     'string',
-  ];
+  ]
   if (typeofA !== typeofB) {
-    return orders.indexOf(typeofA) - orders.indexOf(typeofB);
+    return orders.indexOf(typeofA) - orders.indexOf(typeofB)
   }
 
   if (typeofA !== 'string') {
-    return Number(a) - Number(b);
+    return Number(a) - Number(b)
   }
 
-  return (a as string).localeCompare(b as string);
+  return (a as string).localeCompare(b as string)
 }
 
 export function normalizeInvalidHandler(
   invalidHandler: InvalidHandler,
 ): NormalizedInvalidHandler {
   return (...args) => {
-    const errorMessageOrError = invalidHandler(...args);
+    const errorMessageOrError = invalidHandler(...args)
     return typeof errorMessageOrError === 'string'
-      ? new Error(errorMessageOrError)
-      : /* istanbul ignore next*/ errorMessageOrError;
-  };
+      ? new Error(errorMessageOrError) /* c8 ignore start */
+      : errorMessageOrError /* c8 ignore stop */
+  }
 }
 
 export function normalizeDefaultResult<$Value>(
   result: DefaultResult<$Value>,
 ): NormalizedDefaultResult<$Value> {
-  return result === undefined ? {} : result;
+  return result === undefined ? {} : result
 }
 
 export function normalizeExpectedResult(
   result: ExpectedResult,
 ): NormalizedExpectedResult {
   if (typeof result === 'string') {
-    return { text: result };
+    return { text: result }
   }
 
-  const { text, list } = result;
+  const { text, list } = result
 
   assert(
     (text || list) !== undefined,
     'Unexpected `expected` result, there should be at least one field.',
-  );
+  )
 
   if (!list) {
-    return { text };
+    return { text }
   }
 
   return {
@@ -170,25 +172,25 @@ export function normalizeExpectedResult(
       title: list.title,
       values: list.values.map(normalizeExpectedResult),
     },
-  };
+  }
 }
 
 export function normalizeValidateResult(
   result: ValidateResult,
   value: unknown,
 ): NormalizedValidateResult {
-  return result === true ? true : result === false ? { value } : result;
+  return result === true ? true : result === false ? { value } : result
 }
 
 export function normalizeDeprecatedResult<$Value>(
   result: DeprecatedResult<$Value>,
   value: $Value,
-): NormalizedDeprecatedResult<$Value>;
+): NormalizedDeprecatedResult<$Value>
 export function normalizeDeprecatedResult<$Value>(
   result: DeprecatedResult<$Value>,
   value: $Value,
   doNotNormalizeTrue: true,
-): NormalizedDeprecatedResultWithTrue<$Value>;
+): NormalizedDeprecatedResultWithTrue<$Value>
 export function normalizeDeprecatedResult<$Value>(
   result: DeprecatedResult<$Value>,
   value: $Value,
@@ -199,14 +201,14 @@ export function normalizeDeprecatedResult<$Value>(
   return result === false
     ? false
     : result === true
-      ? doNotNormalizeTrue
-        ? true
-        : [{ value }]
-      : 'value' in result
-        ? [result]
-        : result.length === 0
-          ? false
-          : (result as NotEmptyArray<{ value: $Value }>);
+    ? doNotNormalizeTrue
+      ? true
+      : [{ value }]
+    : 'value' in result
+    ? [result]
+    : result.length === 0
+    ? false
+    : (result as NotEmptyArray<{ value: $Value }>)
 }
 
 export function normalizeTransferResult<$Value>(
@@ -216,8 +218,8 @@ export function normalizeTransferResult<$Value>(
   return typeof result === 'string' || 'key' in result
     ? { from: value, to: result }
     : 'from' in result
-      ? { from: result.from!, to: result.to }
-      : { from: value, to: result.to };
+    ? { from: result.from!, to: result.to }
+    : { from: value, to: result.to }
 }
 
 export function normalizeForwardResult<$Value>(
@@ -227,10 +229,10 @@ export function normalizeForwardResult<$Value>(
   return result === undefined
     ? []
     : Array.isArray(result)
-      ? result.map(transferResult =>
-          normalizeTransferResult(transferResult, value),
-        )
-      : [normalizeTransferResult(result, value)];
+    ? result.map(transferResult =>
+        normalizeTransferResult(transferResult, value),
+      )
+    : [normalizeTransferResult(result, value)]
 }
 
 export function normalizeRedirectResult<$Value>(
@@ -242,17 +244,18 @@ export function normalizeRedirectResult<$Value>(
       ? result.redirect
       : result,
     value,
-  );
+  )
   return redirect.length === 0
     ? { remain: value, redirect }
     : typeof result === 'object' && 'remain' in result
-      ? { remain: result.remain, redirect }
-      : { redirect };
+    ? { remain: result.remain, redirect }
+    : { redirect }
 }
 
 export function assert(isValid: boolean, message: string) {
-  // istanbul ignore if
+  /* c8 ignore start */
   if (!isValid) {
-    throw new Error(message);
+    throw new Error(message)
   }
+  /* c8 ignore stop */
 }
